[backend] chore: diary 도메인에서 사용되는 모든 ID 타입을 int → long 으로 수정

+++ backend/src/main/java/com/example/demo/service/DiaryService.java
@@ -12,7 +12,7 @@ public interface DiaryService {
     void insertDiary(DiaryWriteRequestDTO diaryWriteRequestDTO);
     void updateDiary(long id, DiaryEditRequestDTO diaryData);
     DiaryDetailResponseDTO requestDiaryDetails(long diaryId);
-    void deleteDiary(int diaryId);
-    List<AllTeamDiariesResponse> requestAllTeamDiaries(int userId);
-    List<DiaryModel> requestDiaryId(String diaryTitle, String writtenDate, int writerId);
+    void deleteDiary(long diaryId);
+    List<AllTeamDiariesResponse> requestAllTeamDiaries(long userId);
+    List<DiaryModel> requestDiaryId(String diaryTitle, String writtenDate, long writerId);
 }
